archives/skills.old/cv-manager/scripts/pyproject.toml
[project]
name = "cv-manager-scripts"
version = "0.1.0"
description = "CV Management Scripts"
requires-python = ">=3.10"
dependencies = [
    "google-auth",
    "google-auth-oauthlib",
    "google-auth-httplib2",
    "google-api-python-client",
    "python-dotenv",
    "PyPDF2",
    "openai",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
